+++ src/views/admin-aulas.ejs
@@ -19,7 +19,7 @@
                 </thead>
                 
                 <tbody>
-                    <% aulas.forEach(servico => { %>
+                    <% aulas.forEach(aula => { %>
                     <tr>
                         <td><img src="<%=personal.image %>" alt=""></td>
                         <td><%= aula.nome %> </td>
